requirements.txt
Django
djangorestframework
numpy
opencv-python-headless
